==> Lessons/04_CodingBasics.Rmd <==
---
title: "4: Coding Basics"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Develop familiarity with the form and function of the RStudio interface.
2. Apply basic functionality of R
3. Evaluate how basic practies in R contribute to best management practices for data analysis

## DATA TYPES IN R
R treats objects differently based on their characteristics. For more information, please see: https://www.statmethods.net/input/datatypes.html. 

* **Vectors** 1 dimensional structure that contains elements of the same type (ex: all numbers, all letters).

* **Matrices** 2 dimensional structure that contains elements of the same type.(most of what we will talk about will be 2-d)

* **Arrays** Similar to matrices, but can have more than 2 dimensions. We will not delve into arrays in depth.

* **Lists** Ordered collection of elements that can have different modes.

* **Data Frames** 2 dimensional structure that is more general than a matrix. Columns can have different modes (e.g., numeric and factor). When we import csv files into the R workspace, they will enter as data frames.

Define what each new piece of syntax does below (i.e., fill in blank comments). Note that the R chunk has been divided into sections (# at beginning of line, ---- at end)
```{r}
# Vectors ----
vector1 <- c(1,2,5.3,6,-2,4) # numeric vector
vector1
vector2 <- c("one","two","three") # character vector
vector2
vector3 <- c(TRUE,TRUE,TRUE,FALSE,TRUE,FALSE) #logical vector
vector3

vector1[3] # gives the third item of the vector (the third element of vector1) - matrix subsetting

# Matrices ----
matrix1 <- matrix(1:20, nrow = 5,ncol = 4) # matrix of numbers 1:20 with 5 rows and 4 columns, deafult- automatically counts vertically first
matrix1
matrix2 <- matrix(1:20, nrow = 5, ncol = 4, byrow = TRUE) #counts horizontally first then vertically 
matrix2
matrix3 <- matrix(1:20, nrow = 5, ncol = 4, byrow = TRUE, # return after comma continues the line
                  dimnames = list(c("uno", "dos", "tres", "cuatro", "cinco"), 
                                  c("un", "deux", "trois", "cat"))) #naming rows then columns
matrix3

#indexing
matrix1[4, ] # pull items from row 4
matrix1[ , 3] # pull items from column 3 [row, column]
matrix1[c(12, 14)] # pull element 12 and 14 (need the c or it will think you want it to pull row 12, column 14 - c means "call")
matrix1[c(12:14)] # pull elements 12 through 14
matrix1[2:4, 1:3] # create matrix with only rows 2:4 and columns 1:3
matrix1[,1:3]

cells <- c(1, 26, 24, 68)
rnames <- c("R1", "R2")
cnames <- c("C1", "C2") 
matrix4 <- matrix(cells, nrow = 2, ncol = 2, byrow = TRUE,
  dimnames = list(rnames, cnames)) # 
matrix4
#warning vs error

# Lists ---- 
list1 <- list(name = "Maria", mynumbers = vector1, mymatrix = matrix1, age = 5.3); list1 #naming the object within the code rather than defining them seperately 
list1[[2]]

# Data Frames ----
d <- c(1, 2, 3, 4) # What type of vector? numeric
e <- c("red", "white", "red", NA) # What type of vector? character
f <- c(TRUE, TRUE, TRUE, FALSE) # What type of vector? logical
dataframe1 <- data.frame(d,e,f) #combine these vectors into a matrix
names(dataframe1) <- c("ID","Color","Passed"); View(dataframe1) # View() pulls up the visualization in a seperate window 

dataframe1[1:2] # showing columns 1 and 2 because it is a dataframe not a matrix, if it were a matrix it would give elements 1 and 2
dataframe1[c("ID","Passed")]  
dataframe1$Color 

```
QUESTION: How do the different types of data appear in the Environment tab?

> ANSWER: they have different descriptive statements, lists, matrices and data frames go under data. Vectors go under values.

QUESTION: In the R chunk below, write "dataframe1$". Press `tab` after you type the dollar sign. What happens?

> ANSWER: The first element from the list of data elements appears 

QUESTION: What happens when a comment in R is followed by "----"?

> ANSWER:it becomes a collapsible section

Advanced: Sequential section headers can be created by using at least four -, =, and # characters.

## PACKAGES

The Packages tab in the notebook stores the packages that you have saved in your system. A checkmark next to each package indicates whether the package has been loaded into your current R session. Given that R is an open source software, users can create packages that have specific functionalities, with complicated code "packaged" into a simple commands.

If you want to use a specific package that is not in your libaray already, you need to install it. You can do this in two ways: 

1. Click the install button in the packages tab. Type the package name, which should autocomplete below (case matters). Make sure to check "intall dependencies," which will also install packages that your new package uses. 

2. Type `install.packages("packagename")` into your R chunk or console (do this in the console so that you dont have the install.packages function in your script a) because you only need to do this once b)it will cause knitting problems). It will then appear in your packages list. You only need to do this once. 

If a package is already installed, you will need to load it every session. You can do this in two ways: 

1. Click the box next to the package name in the Packages tab. 

2. Type `library(packagename)` into your R chunk or console.

2a. The command `require(packagename)` will also load a package, but it will not give any error or warning messages if there is an issue.

**Tips and troubleshooting**

* You may be asked to restart R when installing or updating packages. Feel free to say no, as this will obviously slow your progress. However, if the functionality of your new package isn't working properly, try restarting R as a first step. 

* If asked "Do you want to install from sources the packages which needs compilation?", type `yes` into the console. 

* You should only install packages once on your machine. If you store `install.packages` in your R chunks/scripts, comment these lines out, as below. 

* Update your packages regularly! go to packages tab and hit update


```{r}
# We will use the packages dplyr and ggplot2 regularly. 
#install.packages("dplyr") # comment out install commands, use only when needed and re-comment
#install.packages("ggplot2")

library(dplyr)
library(ggplot2)

# Some packages are umbrellas under which other packages are loaded
#install.packages("tidyverse")
library(tidyverse)
```
What happens in the console when you load a package?

> ANSWER: 

## FUNCTIONS

You've had some practice with functions with the simple commands you've entered in this lesson and the one previous. The basic form of a function is `functionname()`, and the packages we will use in this class will use these basic forms. However, there may be situations when you will want to create your own function. Below is a description of how to write functions through the metaphor of creating a recipe. Credit for this goes to Isabella R. Ghement (@IsabellaGhement on Twitter). 

Writing a function is like writing a recipe. Your function will need a recipe name (functionname). Your recipe ingredients will go inside the parentheses. The recipe steps and end product go inside the curly brackets.
```{r}
functionname <- function(){
  
}
```

A single ingredient recipe: 
```{r}
# Write the recipe
recipe1 <- function(x){
  mix <- x*2
  return(mix)
}

# Bake the recipe
simplemeal <- recipe1(5)

# Serve the recipe
simplemeal
```

Two single ingredient recipes, baked at the same time: 
```{r}
recipe2 <- function(x){
  mix1 <- x*2
  mix2 <- x/2
  return(list(mix1 = mix1, #comma indicates we continue onto the next line
              mix2 = mix2))
}

doublesimplemeal <- recipe2(6)
doublesimplemeal
```
Two double ingredient recipes, baked at the same time: 
```{r}
recipe3 <- function(x, f){
  mix1 <- x*f
  mix2 <- x/f
  return(list(mix1 = mix1, #comma indicates we continue onto the next line
              mix2 = mix2))
}

doublecomplexmeal <- recipe3(x = 5, f = 2)
doublecomplexmeal
doublecomplexmeal$mix1
```
Make a recipe based on the ingredients you have
```{r}
recipe4 <- function(x) {
  if(x < 3) {x*2} 
  else {x/2}
}

recipe5 <- function(x) {
  if(x < 3) {x*2} 
  else if (x > 3) {x/2}
  else {x}
}
meal <- recipe4(4); meal
meal2 <- recipe4(2); meal2
meal3 <- recipe5(3); meal3

recipe6 <- function(x){
  ifelse(x<3, x*2, x/2)
  
}

meal4 <- recipe6(4); meal4
meal5 <- recipe6(2); meal5
```
